+++ src/plugins/rehype/external-links.ts
@@ -1,13 +1,13 @@
 import rehypeExternalLinks, { type Options } from "rehype-external-links";
 import type { Element } from "hast";
 
+import { externalLinkArrow } from "../shared";
+
 function hasImgChild(node: Element): boolean {
 	return node.children.some(
 		(child) => child.type === "element" && child.tagName === "img",
 	);
 }
-
-export const externalLinkArrow = " ↗";
 
 export const rehypeExternalLinksOptions = {
 	content: (element) => {
